// frontend/src/app/generated/client/fn/admin/api-admin-audit-get-json.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { PagedResultOfAuditEntryDto } from '../../models/paged-result-of-audit-entry-dto';

export interface ApiAdminAuditGet$Json$Params {
  entityType?: string;
  entityId?: string;
  action?: string;
  userId?: string;
  from?: string;
  to?: string;
  page?: (number | string);
  pageSize?: (number | string);
  sortBy?: string;
  sortDescending?: boolean;
}

export function apiAdminAuditGet$Json(http: HttpClient, rootUrl: string, params?: ApiAdminAuditGet$Json$Params, context?: HttpContext): Observable<StrictHttpResponse<PagedResultOfAuditEntryDto>> {
  const rb = new RequestBuilder(rootUrl, apiAdminAuditGet$Json.PATH, 'get');
  if (params) {
    rb.query('entityType', params.entityType, {});
    rb.query('entityId', params.entityId, {});
    rb.query('action', params.action, {});
    rb.query('userId', params.userId, {});
    rb.query('from', params.from, {});
    rb.query('to', params.to, {});
    rb.query('page', params.page, {});
    rb.query('pageSize', params.pageSize, {});
    rb.query('sortBy', params.sortBy, {});
    rb.query('sortDescending', params.sortDescending, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'text/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<PagedResultOfAuditEntryDto>;
    })
  );
}

apiAdminAuditGet$Json.PATH = '/api/Admin/audit';