// frontend/src/app/generated/client/models/audit-entry-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AuditEntryDto {
  action: string;
  changesJson: (string | null);
  entityId: string;
  entityType: string;
  id: (number | string);
  occurredAt: string;
  userEmail: (string | null);
  userId: (string | null);
}